requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.17
typing-extensions==4.7.0
Werkzeug==2.3.6
zipp==3.15.0
